% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-data.R
\name{distribute_unassigned_nyt}
\alias{distribute_unassigned_nyt}
\title{Distribute cases and deaths from "Unknown" counties to mappable counties}
\usage{
distribute_unassigned_nyt(
  df,
  probdf,
  distribute_unknowns = FALSE,
  skip_assignment = NULL,
  na_fill = FALSE,
  seed = NULL
)
}
\arguments{
\item{df}{data.frame of county level covid-19 data from NYT}

\item{probdf}{data.frame. Data to be used as probabilities for assigning observations.
required to have county fips codes.}

\item{distribute_unknowns}{logical. If \code{TRUE} distribute all "Unknown" values for each state}

\item{skip_assignment}{character vector. A vector of state fips codes to skip in the distribution
of unknown cases.}

\item{na_fill}{logical. default is \code{FALSE}}

\item{seed}{integer. If set, used in \code{\link[=set.seed]{set.seed()}}.Default is \code{NULL}}
}
\value{
data frame
}
\description{
This function is for handling the specific issues with the data being reported
by the NYT in this \href{https://github.com/nytimes/covid-19-data}{repo},
so there are some hard coded things, if the issues change, then this might no longer work.
}
