% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-data.R
\name{distribute_unassigned_state}
\alias{distribute_unassigned_state}
\title{Distribute unassigned values to counties based on multinomial probabilities}
\usage{
distribute_unassigned_state(
  df,
  location_to_distribute,
  assign_values_to = NULL,
  na_fill = FALSE,
  seed = 1998
)
}
\arguments{
\item{df}{data.frame of a single state}

\item{location_to_distribute}{unique id of location whose values will be distributed}

\item{assign_values_to}{unqiued ids of locations that will be recieving the unassigned values.
vector of ids or if \code{NULL} all non \code{NA} fips values are used.
Default is \code{NULL}}

\item{na_fill}{For days in which there are no reported data, how to fill in the \code{NA}s}

\item{seed}{integer. If set, used in \code{\link[=set.seed]{set.seed()}}.Default is \code{1998}}
}
\value{
data frame
}
\description{
This function is intented to be used under the hood of wrapper functions that
distributes unassigned values for a specific dataset. For example,
\code{\link[=distribute_unassigned_nyt]{distribute_unassigned_nyt()}}.
}
