% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-data.R
\name{format_nyt}
\alias{format_nyt}
\title{Format New York Times COVID-19 County Level Data}
\usage{
format_nyt(
  df,
  distribute_unknowns = TRUE,
  skip_assignment = NULL,
  na_fill = FALSE,
  seed = 1998
)
}
\arguments{
\item{df}{the result of calling \code{\link[=get_nyt]{get_nyt()}}}

\item{distribute_unknowns}{logical. Distribute cases in "Unknown" counties to
actual counties. Default is \code{TRUE}}

\item{skip_assignment}{character vector. A vector of state fips codes to skip in the distribution
of unknown cases.}

\item{na_fill}{logical. Default is \code{FALSE}}

\item{seed}{integer. If set, used in \code{\link[=set.seed]{set.seed()}}.Default is \code{1998}}
}
\value{
a data frame with additionally formatted data
}
\description{
This function is for formatting specific issues with the data being reported
by the NYT county data in this \href{https://github.com/nytimes/covid-19-data}{repo}
}
\examples{
library(covidmap)

df_raw <- get_nyt()

df <- format_nyt(df_raw)
}
