% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility-data.R
\name{get_descartes_mobility}
\alias{get_descartes_mobility}
\title{Get Latest Mobility Data from Google}
\usage{
get_descartes_mobility(admin = c("county", "state", "all"))
}
\arguments{
\item{admin}{character. One of county, state, or all.}
}
\value{
df
}
\description{
\href{https://descarteslabs.com/}{Descartes Labs} is releasing mobility statistics (representing the distance a
typical member of a given population moves in a day) at the US admin1 (state)
and admin2 (county) level. A technical report describing the motivation behind
this work with methodology and definitions is available \href{descarteslabs.com/mobility-v097}{here (pdf)}.
}
\details{
Note: Data for \code{2020-04-20} did not meet quality control standards, and was not released.

From the methodology paper....

These statistics include the mean, median and quartiles. For the results reported here,
we focus on the median of the max-distance mobility Mmax, which we designate \code{m50}.
We further define a normalized mobility index \code{m50_index = 100 * (m50 / m50_norm)} where
\code{m50_norm} is a "normal" value of \code{m50} in a region, defined as the median \code{m50} in that
region during a designated earlier time period. We use the median weekday value of \code{m50}
between the dates of \code{2020-02-17} and \code{2020-03-07} as \code{m50_norm} to investigate
COVID-19 related changes in the US. Note that \code{m50} has dimensions of kilometers,
while \code{m50_index} is dimensionless. \code{m50_index} is easily transformed to a percentage
change from baseline behavior via \code{m50_pct_change = m50_index - 100}
}
\references{
Warren & Skillman 2020. Mobility Changes in Response to COVID-19. \url{descarteslabs.com/mobility-v097}
}
