% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility-data.R
\name{get_mobility}
\alias{get_mobility}
\title{Get Latest Mobility Data from Descartes Labs}
\usage{
get_mobility(admin = c("county", "state", "all"))
}
\arguments{
\item{admin}{character. county, state, or all}
}
\value{
df
}
\description{
Get Latest Mobility Data from Descartes Labs
}
