% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-data.R
\name{get_nyt}
\alias{get_nyt}
\title{Get Latest COVID-19 Data from New York Times Database}
\usage{
get_nyt(admin = c("county", "state"))
}
\arguments{
\item{admin}{character. county level data or state level}
}
\value{
df
}
\description{
Get Latest COVID-19 Data from New York Times Database
}
